requirements.txt
langchain-groq
langchain
python-dotenv
pypdf
chromadb
langchain-community
flask